requirements.txt
# ============================================================
# Veille Juridique Maroc – Python dependencies
# Install: pip install -r requirements.txt
# Python 3.11+ recommended
# ============================================================

# ---- Scraping ----
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0                 # faster HTML parser for bs4
html5lib>=1.1               # fallback parser

# ---- AI / Claude ----
anthropic>=0.28.0

# ---- Database ----
sqlalchemy>=2.0.0
alembic>=1.13.0             # optional: DB migrations

# ---- Scheduler ----
APScheduler>=3.10.0

# ---- Dashboard ----
streamlit>=1.35.0
pandas>=2.2.0

# ---- Configuration & env ----
python-dotenv>=1.0.0
pyyaml>=6.0.1

# ---- Optional: PostgreSQL driver ----
# psycopg2-binary>=2.9.9

# ---- Development / testing ----
pytest>=8.0.0
pytest-cov>=4.1.0
responses>=0.25.0            # mock HTTP requests in tests
